+++ src/lib/paypalUtils.js
